requirements.txt
asgiref==3.3.1
Django==3.1.7
django-cors-headers==3.7.0
djangorestframework==3.12.2
pip==21.0.1
pytz==2021.1
setuptools==49.2.1
sqlparse==0.4.1